fix: extract _merge_queue_group_ci submodule to satisfy 500-line budget and add dropped_merge_group_ci to docstring
- Move _query_merge_group_ci, _QUERY, validation block, and 3 mutation
  strings from merge_queue.py (570→461 lines) to new private submodule
  _merge_queue_group_ci.py; re-export all symbols so existing tests and
  monkeypatches work unchanged
- Add dropped_merge_group_ci to wait_for_merge_queue docstring to satisfy
  test_pr_state_docstring_documents_all_members
