+++ src/autoskillit/server/tools_ci_merge_queue.py
@@ -218,20 +218,22 @@ async def wait_for_merge_queue(
         JSON: {
             "success": bool,
             "pr_state": "merged"|"ejected"|"ejected_ci_failure"|"stalled"|
-                        "dropped_healthy"|"not_enrolled"|"timeout"|"error",
+                        "dropped_healthy"|"dropped_merge_group_ci"|
+                        "not_enrolled"|"timeout"|"error",
             "reason": str,
             "stall_retries_attempted": int,
         }
 
         pr_state values:
-          merged           — PR successfully merged through the queue.
-          ejected          — PR removed from queue (conflict or other non-CI reason).
-          ejected_ci_failure — PR removed from queue because CI checks failed.
-          stalled          — PR stuck in queue; max stall retries exhausted.
-          dropped_healthy  — auto-merge disabled on a PR with no CI/conflict issues.
-          not_enrolled     — PR was never enrolled in the merge queue.
-          timeout          — polling budget exhausted before a terminal state was reached.
-          error            — watcher raised an unexpected exception.
+          merged                — PR successfully merged through the queue.
+          ejected               — PR removed from queue (conflict or other non-CI reason).
+          ejected_ci_failure    — PR removed from queue because CI checks failed.
+          stalled               — PR stuck in queue; max stall retries exhausted.
+          dropped_healthy       — auto-merge disabled on a PR with no CI/conflict issues.
+          dropped_merge_group_ci — PR ejected due to merge-group CI failure (PR-branch CI clean).
+          not_enrolled          — PR was never enrolled in the merge queue.
+          timeout               — polling budget exhausted before a terminal state was reached.
+          error                 — watcher raised an unexpected exception.
 
     Never raises.
     """
